+++ src/commands/plan.rs
@@ -762,6 +762,7 @@ impl Plan {
                 | SystemVerilogSymbol::Package(_)
                 | SystemVerilogSymbol::Primitive(_)
                 | SystemVerilogSymbol::Checker(_)
+                | SystemVerilogSymbol::Program(_)
                 | SystemVerilogSymbol::Interface(_) => {
                     component_pairs.insert(
                         LangIdentifier::Verilog(name.unwrap().clone()),